Clean up App.kt
